Extract secret detection into secret_redaction and regex_dfas crates
Move the secret regexes, the default secret patterns, and the text
detection and redaction helpers out of warp_terminal and app into a new
secret_redaction crate. Move the regex DFA search core out of
warp_terminal's find module into a new regex_dfas crate, and keep the
grid search loop in warp_terminal as GridHandler methods. Rename
SecretRange to StringRange and move it into string-offset.
