+++ app/src/terminal/model/secrets.rs
@@ -1,12 +1,12 @@
+use string_offset::StringRange;
 pub use warp_terminal::model::secrets::*;
 use warpui::EntityId;
-use warpui::elements::SecretRange;
 
 use crate::ai::blocklist::block::TextLocation;
 #[derive(Clone, Debug)]
 pub struct RichContentSecretTooltipInfo {
     pub secret: String,
-    pub secret_range: SecretRange,
+    pub secret_range: StringRange,
     pub location: TextLocation,
     pub is_obfuscated: bool,
     pub position_id: String,
