+++ crates/warpui_core/src/elements/gui/formatted_text_element_tests.rs
@@ -8,11 +8,11 @@ use markdown_parser::{FormattedText, FormattedTextFragment, FormattedTextLine};
 use pathfinder_color::ColorU;
 use pathfinder_geometry::rect::RectF;
 use pathfinder_geometry::vector::vec2f;
-use string_offset::ByteOffset;
+use string_offset::{ByteOffset, StringRange};
 
 use super::{
     FormattedTextElement, FrameMouseHandlers, HeadingFontSizeMultipliers, HighlightedHyperlink,
-    HyperlinkSupport, LaidOutTextFrame, SecretRange, apply_secret_replacements,
+    HyperlinkSupport, LaidOutTextFrame, apply_secret_replacements,
 };
 use crate::elements::{Element, PartialClickableElement, Point, SelectableElement, ZIndex};
 use crate::event::DispatchedEvent;
@@ -101,8 +101,8 @@ fn test_custom_heading_font_size_multipliers() {
     );
 }
 
-fn sr(char_start: usize, char_end: usize, byte_start: usize, byte_end: usize) -> SecretRange {
-    SecretRange {
+fn sr(char_start: usize, char_end: usize, byte_start: usize, byte_end: usize) -> StringRange {
+    StringRange {
         char_range: char_start..char_end,
         byte_range: byte_start..byte_end,
     }
