+++ crates/warpui_core/src/elements/gui/mod.rs
@@ -73,6 +73,7 @@ pub use selectable_area::*;
 pub use shared_scrollbar::*;
 pub use size_constraint_switch::*;
 pub use stack::*;
+use string_offset::StringRange;
 pub use table::{
     RowBackground, Table, TableColumnWidth, TableConfig, TableHeader, TableState, TableStateHandle,
     TableVerticalSizing,
@@ -703,23 +704,6 @@ impl HoverableCharRange {
     }
 }
 
-/// SecretRange is used to store both the char range and byte range of a secret.
-/// We need to do this since several APIs e.g. hover/click APIs, use char ranges,
-/// whereas text-related APIs e.g. Regex and replace_range, use byte ranges.
-#[derive(Debug, Eq, PartialEq, Hash, Clone)]
-pub struct SecretRange {
-    pub char_range: Range<usize>,
-    pub byte_range: Range<usize>,
-}
-
-impl SecretRange {
-    /// Extends the current range to include the provided range.
-    pub fn extend_range_end(&mut self, other: &SecretRange) {
-        self.char_range.end = self.char_range.end.max(other.char_range.end);
-        self.byte_range.end = self.byte_range.end.max(other.byte_range.end);
-    }
-}
-
 pub trait PartialClickableElement {
     /// clickable_char_ranges is the vector of char ranges that the caller can
     /// specify, where the callback will be called if any character in one of
@@ -745,7 +729,7 @@ pub trait PartialClickableElement {
         F: 'static + FnMut(bool, &mut EventContext, &AppContext);
 
     /// Replace in the given range of the text with the replacement text.
-    fn replace_text_range(&mut self, range: SecretRange, replacement: Cow<'static, str>);
+    fn replace_text_range(&mut self, range: StringRange, replacement: Cow<'static, str>);
 }
 
 /// An element that can be selected, for use with the SelectableArea element.
